move fluid.layer.py_func to paddle.static.nn.common.py_func
