+++ python/paddle/static/__init__.py
@@ -16,6 +16,9 @@
 from . import amp  # noqa: F401
 from . import sparsity  # noqa: F401
 from . import nn  # noqa: F401
+
+from .nn.common import py_func  # noqa: F401
+
 from .io import save_inference_model  # noqa: F401
 from .io import load_inference_model  # noqa: F401
 from .io import deserialize_persistables  # noqa: F401
@@ -53,7 +56,6 @@ from ..fluid.framework import Variable  # noqa: F401
 from ..fluid.framework import ipu_shard_guard  # noqa: F401
 from ..fluid.framework import set_ipu_shard  # noqa: F401
 from ..fluid.layers.control_flow import Print  # noqa: F401
-from ..fluid.layers.nn import py_func  # noqa: F401
 from ..fluid.parallel_executor import ParallelExecutor  # noqa: F401
 from ..fluid.param_attr import WeightNormParamAttr  # noqa: F401
 from ..fluid.optimizer import ExponentialMovingAverage  # noqa: F401
@@ -61,7 +63,6 @@ from ..fluid.io import save  # noqa: F401
 from ..fluid.io import load  # noqa: F401
 from ..fluid.io import load_program_state  # noqa: F401
 from ..fluid.io import set_program_state  # noqa: F401
-
 from ..fluid.io import load_vars  # noqa: F401
 from ..fluid.io import save_vars  # noqa: F401
 from ..fluid.io import batch  # noqa: F401
